+++ sdk/formrecognizer/ai-form-recognizer/test/public/node/formrecognizerclient.spec.ts
@@ -26,7 +26,7 @@ describe("FormRecognizerClient NodeJS only", () => {
 
   afterEach(async function() {
     if (recorder) {
-      recorder.stop();
+      await recorder.stop();
     }
   });
 
@@ -185,7 +185,7 @@ describe("[AAD] FormRecognizerClient NodeJS only", () => {
 
   afterEach(async function() {
     if (recorder) {
-      recorder.stop();
+      await recorder.stop();
     }
   });