+++ sdk/metricsadvisor/ai-metrics-advisor/test/adminclient.spec.ts
@@ -32,7 +32,7 @@ describe("MetricsAdvisorAdministrationClient", () => {
 
   afterEach(async function() {
     if (recorder) {
-      recorder.stop();
+      await recorder.stop();
     }
   });